backend/requirements.txt
# FastAPI web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# YouTube transcript extraction (latest version)
youtube-transcript-api>=0.6.2

# Google Gemini AI
google-generativeai==0.3.1

# Deepgram SDK (fallback for videos without captions)
deepgram-sdk==3.2.0

# Environment variables
python-dotenv==1.0.0

# CORS middleware
python-multipart==0.0.6

# HTTP client for API calls (updated for deepgram-sdk 3.2.0 compatibility)
httpx>=0.25.2

# Authentication
pyjwt==2.8.0

# Email validation
pydantic[email]>=2.0.0

# Supabase client with pinned gotrue to fix proxy error
supabase==2.4.0
gotrue==1.3.0

# Pinecone embeddings (free tier for semantic search)
# Note: Package renamed from pinecone-client to pinecone
pinecone>=3.0.0

# NumPy for fast similarity calculations
numpy>=1.24.0

# BM25 lexical retrieval (for A/B testing vs embeddings)
rank-bm25>=0.2.2

# Background task scheduler
apscheduler>=3.10.0
